WIP: Porting to x86_64: System modules
0f4486d404e8d61e86cfb6053b299d71cdef2cb9a6935145ae5ab7a67159f525

- 64-bit System module support
- Linter now checks for full license

+++ include/sysmods/interface.h
@@ -23,8 +23,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 typedef struct {
     char name[32];
-    uint32_t init_offset; // relative offset of the function
-    uint32_t exit_offset;
+    uint64_t init_offset; // relative offset of the function
+    uint64_t exit_offset;
 } __attribute__((packed)) sysmod_t;
 
 typedef struct {
